fix(tires): unblock prod backfill and fleet recalculate ops
- Bootstrap Nest via AppModule.forRootAsync() in tire backfill ops scripts
- Cast advisory lock keys to int for PostgreSQL pg_advisory_xact_lock
- Remove erroneous throw on static blockedWearDerivations list in wear model
- Add tire-fleet-recalculate.ts ops script
- Archive post-ops integrity audit snapshot

+++ backend/src/modules/vehicle-intelligence/tires/tire-trip-usage-replay.ts
@@ -98,7 +98,7 @@ export async function acquireTripUsageAdvisoryLock(
   tireSetupId: string,
 ): Promise<void> {
   const [k1, k2] = pgAdvisoryLockKeys(advisoryLockSeed(tripId, tireSetupId));
-  await tx.$executeRaw`SELECT pg_advisory_xact_lock(${k1}, ${k2})`;
+  await tx.$executeRaw`SELECT pg_advisory_xact_lock((${k1})::int, (${k2})::int)`;
 }
 
 type LedgerAggregateClient = {
